Getting complete rid of Java API
