--- api/src/main/scala/com/avast/clients/rabbitmq/javaapi/RabbitMQConsumer.scala
@@ -1,3 +1,0 @@
-package com.avast.clients.rabbitmq.javaapi
-
-trait RabbitMQConsumer extends AutoCloseable
