--- core/src/main/scala/com/avast/clients/rabbitmq/javaapi/DefaultRabbitMQConsumer.scala
@@ -1,3 +1,0 @@
-package com.avast.clients.rabbitmq.javaapi
-
-abstract class DefaultRabbitMQConsumer private[javaapi] () extends RabbitMQConsumer
