--- core/src/main/scala/com/avast/clients/rabbitmq/javaapi/DefaultRabbitMQPullConsumer.scala
@@ -1,25 +1,0 @@
-package com.avast.clients.rabbitmq.javaapi
-
-import java.util.concurrent.CompletableFuture
-
-import com.avast.bytes.Bytes
-import com.avast.clients.rabbitmq.api.{PullResult => ScalaResult, RabbitMQPullConsumer => ScalaPullConsumer}
-import com.avast.clients.rabbitmq.javaapi.JavaConverters._
-import com.typesafe.scalalogging.StrictLogging
-
-import scala.concurrent.{ExecutionContext, Future}
-
-private abstract class DefaultRabbitMQPullConsumer private[javaapi] (scalaConsumer: ScalaPullConsumer[Future, Bytes])(
-    implicit ec: ExecutionContext)
-    extends RabbitMQPullConsumer
-    with StrictLogging {
-  override def pull(): CompletableFuture[PullResult] = {
-    scalaConsumer
-      .pull()
-      .map[PullResult] {
-        case ScalaResult.Ok(dwh) => new PullResult.Ok(dwh.asJava)
-        case ScalaResult.EmptyQueue => PullResult.EmptyQueue
-      }
-      .asJava
-  }
-}
